Fix minor gcc warnings
(-Wsign-compare, -Wunused-parameter, -Wreorder, -Wunused-variable)


Former-commit-id: b727940e1953eed11054aa7d7d4f8eae3eeb37f1

+++ modules/sofa/component/container/MechanicalObject.inl
@@ -1468,7 +1468,7 @@ Data<typename MechanicalObject<DataTypes>::VecCoord>* MechanicalObject<DataTypes
             vectorsCoord[v.index]->beginEdit()->reserve(f_reserve.getValue());
             vectorsCoord[v.index]->endEdit();
         }
-        if( vectorsCoord[v.index]->getValue().size() != getSize() )
+        if(vectorsCoord[v.index]->getValue().size() != (size_t)getSize())
         {
             vectorsCoord[v.index]->beginEdit()->resize( getSize() );
             vectorsCoord[v.index]->endEdit();
@@ -1532,7 +1532,7 @@ Data<typename MechanicalObject<DataTypes>::VecDeriv>* MechanicalObject<DataTypes
             vectorsDeriv[v.index]->beginEdit()->reserve(f_reserve.getValue());
             vectorsDeriv[v.index]->endEdit();
         }
-        if( vectorsDeriv[v.index]->getValue().size() != getSize() )
+        if(vectorsDeriv[v.index]->getValue().size() != (size_t)getSize())
         {
             vectorsDeriv[v.index]->beginEdit()->resize( getSize() );
             vectorsDeriv[v.index]->endEdit();
